+++ verik-compiler/src/test/kotlin/io/verik/compiler/transform/upper/InjectedExpressionTransformerStageTest.kt
@@ -17,6 +17,7 @@
 package io.verik.compiler.transform.upper
 
 import io.verik.compiler.test.BaseTest
+import io.verik.compiler.test.findDeclaration
 import io.verik.compiler.test.findExpression
 import org.junit.jupiter.api.Test
 
@@ -48,5 +49,19 @@ internal class InjectedExpressionTransformerStageTest : BaseTest() {
             InjectedExpressionTransformerStage::class,
             "InjectedExpression(Void, [abc])"
         ) { it.findExpression("f") }
+    }
+
+    @Test
+    fun `injected expression constraint`() {
+        driveElementTest(
+            """
+                class C : Class() {
+                    @Cons
+                    val c = c("1")
+                }
+            """.trimIndent(),
+            InjectedExpressionTransformerStage::class,
+            "Constraint(c, BlockExpression(Unit, [InjectedExpression(Void, [1])]))"
+        ) { it.findDeclaration("c") }
     }
 }