+++ verik-core/src/main/kotlin/io/verik/core/AssociativeArray.kt
@@ -14,11 +14,26 @@
  * limitations under the License.
  */
 
-@file:Suppress("unused")
+@file:Suppress("unused", "UNUSED_PARAMETER")
 
 package io.verik.core
 
 /**
  * An associative array of elements with key [K] and value [V]. It corresponds to a SystemVerilog associative array.
  */
-class AssociativeArray<K, V> private constructor()
+class AssociativeArray<K, V> private constructor() {
+
+    /**
+     * Set the [value] corresponding to [key].
+     */
+    operator fun set(key: K, value: V) {
+        throw VerikException()
+    }
+
+    /**
+     * Get the value corresponding to [key].
+     */
+    operator fun get(key: K): V {
+        throw VerikException()
+    }
+}