+++ verik-kotlin/src/main/kotlin/io/verik/kotlin/String.kt
@@ -14,11 +14,19 @@
  * limitations under the License.
  */
 
-@file:Suppress("unused")
+@file:Suppress("unused", "UNUSED_PARAMETER")
 
 package io.verik.kotlin
 
 /**
  * The String class represents character strings.
  */
-class String private constructor()
+class String private constructor() {
+
+    /**
+     * Returns a string obtained by concatenating this string with the string representation of [other].
+     */
+    operator fun plus(other: Any?): String {
+        throw VerikException()
+    }
+}